Extract the HtmlProvider from the Jinja2 API into its own module.
