+++ betty/plugin/trees/__init__.py
@@ -9,7 +9,7 @@ from typing import Optional, List, Tuple, Type, Callable, Iterable, Any
 from betty.event import Event
 from betty.fs import DirectoryBackup
 from betty.generate import PostGenerateEvent
-from betty.jinja2 import HtmlProvider
+from betty.html import HtmlProvider
 from betty.plugin import Plugin, NO_CONFIGURATION
 from betty.site import Site
